explicit type
git-svn-id: svn+ssh://howdy.physics.nyu.edu/usr/local/svn/kcorrect/trunk@61 75dff457-8751-0410-ab1a-cd701654ce6b

+++ src/k_midpnt.c
@@ -15,7 +15,7 @@ double k_midpnt(double (*func)(double), double a, double b, IDL_LONG n)
 	} else {
 		for(it=1,j=1;j<n-1;j++) it *= 3;
 		tnm=it;
-		del=(b-a)/(3.0*tnm);
+		del=(b-a)/(3.0*(double)tnm);
 		ddel=del+del;
 		x=a+0.5*del;
 		sum=0.0;
@@ -25,7 +25,7 @@ double k_midpnt(double (*func)(double), double a, double b, IDL_LONG n)
 			sum += FUNC(x);
 			x += del;
 		}
-		s=(s+(b-a)*sum/tnm)/3.0;
+		s=(s+(b-a)*sum/(double)tnm)/3.0;
 		return s;
 	}
 }
